Changes: A lot of stuff... lost track T___T - Added Knob.kt (to evolve into PluginElement.kt later) - Refactored VolumeKnob (now in state of VolumeKnobNew) --- now using test Knob.kt data instead of hardcoded values (tho it's partially in LocalPluginsDataProvider) --- made it's size static (at least for now) --- now labels should be always visible --- added measurement name to the value indicator - Refactored PluginCard a bit --- Added FlowRow for basic auto knob placement --- tweaked Plugin elements loading a bit --- tweaked Plugin name overflowing for a static header height (now there might be an issue is with very thin, single-line plugins :0) - probably something else I've missed...
