+++ app/src/main/java/com/mukuro/pedalboard/data/Knob.kt
@@ -1,0 +1,9 @@
+package com.mukuro.pedalboard.data
+
+data class Knob(
+    val id: Long,
+    val name: String,
+    val startPoint: Float,
+    val endPoint: Float,
+    val measure: String
+)
