+++ app/src/main/java/com/mukuro/pedalboard/data/Plugin.kt
@@ -7,5 +7,6 @@ data class Plugin(
     val id: Long,
     val type: PluginType = PluginType.ALL,
     val name: String,
-    val aspectRatio: Float
+    val aspectRatio: Float,
+    val elements: List<Knob?>
 )
